// src/lib/replay/analyzers/types/GameType.ts
export enum GameType {
    UNKNOWN = 0,
    FLAG_SOLO_QUEUE = 1,
    FLAG_COOP = 1 << 1,
    FLAG_PVP = 1 << 2,
    FLAG_DRAFT = 1 << 3,
    FLAG_RANKED = 1 << 4,
    MODE_PRACTICE = 1 << 5,
    MODE_AI = 1 << 6,
    MODE_BRAWL = 1 << 7,
    MODE_QM = 1 << 8,
    MODE_UR = 1 << 9,
    MODE_HL = 1 << 10,
    MODE_TL = 1 << 11,
    MODE_CUSTOM = 1 << 12,

    PRACTICE = MODE_PRACTICE | FLAG_SOLO_QUEUE,
    SOLO_AI = MODE_AI | FLAG_SOLO_QUEUE,
    COOP_AI = MODE_AI | FLAG_COOP,
    CUSTOM = MODE_CUSTOM | FLAG_PVP,
    CUSTOM_DRAFT = MODE_CUSTOM | FLAG_PVP | FLAG_DRAFT,
    BRAWL = MODE_BRAWL | FLAG_PVP,
    QUICK_MATCH = MODE_QM | FLAG_PVP,
    UNRANKED_DRAFT = MODE_UR | FLAG_PVP | FLAG_DRAFT,
    HERO_LEAGUE = MODE_HL | FLAG_PVP | FLAG_DRAFT | FLAG_RANKED | FLAG_SOLO_QUEUE,
    TEAM_LEAGUE = MODE_TL | FLAG_PVP | FLAG_DRAFT | FLAG_RANKED
}